Rename data to downloader to accurately describe functionality
